+++ src/main/vue/src/views/meeting/Detail.vue
@@ -6,14 +6,14 @@ import { useRoute } from "vue-router";
 import api from "@/api"; //index.js
 import Article from "@/components/meeting/Article.vue";
 import Participants from "@/components/meeting/Participants.vue";
-import Comments from "@/components/meeting/Comments.vue";
+import CommentList from "@/components/meeting/CommentList.vue";
 
 export default {
   name: "MeetingDetail",
   components: {
     Article,
     Participants,
-    Comments
+    CommentList,
   },
   setup() {
     const state = reactive({
@@ -43,7 +43,9 @@ export default {
   <div class="content-wrapper">
   <Article :article="state.detail"/>
   <Participants :detail="state.detail"/> 
-  <Comments :detail="state.detail"/> 
+  <CommentList :detail="state.detail"/> 
+
+
   </div>
 </template>
 <style scoped>
